+++ apps/cowswap-frontend/src/modules/tradeQuote/hooks/useTradeQuote.ts
@@ -1,7 +1,6 @@
 import { useAtomValue } from 'jotai'
 
-import { currentTradeQuoteAtom } from '../state/tradeQuoteAtom'
-import { TradeQuoteState } from '../state/tradeQuoteAtom'
+import { currentTradeQuoteAtom, TradeQuoteState } from '../state/tradeQuoteAtom'
 
 export function useTradeQuote(): TradeQuoteState {
   return useAtomValue(currentTradeQuoteAtom)